Fix cargo clippy issues (#6)
* Fix all clippy issues

* Format

* Add system dependency installation to simplified CI clippy job

The --all-features flag pulls in crates that require system libraries
like libx11-dev, libfontconfig-dev, etc. The ubuntu-latest runner
doesn't include these by default, so we need to run ./script/linux
to install them before running clippy.

---------

+++ crates/editor/src/selection.rs
@@ -762,10 +762,10 @@ impl Editor {
                                     let node_start_byte = node.start_byte();
                                     let node_start = node_range.start.0;
                                     let content_start = MultiBufferOffset(
-                                        node_start + (start_byte - node_start_byte) as usize,
+                                        node_start + (start_byte - node_start_byte),
                                     );
                                     let content_end = MultiBufferOffset(
-                                        node_start + (end_byte - node_start_byte) as usize,
+                                        node_start + (end_byte - node_start_byte),
                                     );
                                     let content_range = content_start..content_end;
                                     let current_length = (new_range.end.0 as isize
